Saucenao搜索功能，整理了一下代码

+++ Theresa3rd-Bot/Model/Config/SetuConfig.cs
@@ -16,7 +16,7 @@ namespace Theresa3rd_Bot.Model.Config
 
         public string EnableTagCommand { get; set; }
 
-        public int MaxDaily { get; set; }
+        public long MaxDaily { get; set; }
 
         public int RevokeInterval { get; set; }
